chapter1
ex1 A working Java program is all scrambled up on the fridge.  Can you rearrange the code snippets to make a working Java program that produces the output listed below?  Some of the curly braces fell on the floor and they were too small to pick up, so feel free to add as many of those as you need!